adjusts mx_spawn to omi_spawn_point
