+++ Packages/thirdroom-unity-exporter/Runtime/Scripts/OMI_spawn_point.meta
