+++ Packages/thirdroom-unity-exporter/Runtime/Scripts/OMI_spawn_point/OMI_SpawnPointExtension.cs.meta
